+++ Components/Twitch/LuaGlue.cs
@@ -1,30 +1,24 @@
 ﻿using NLua;
+using Slipstream.Components;
+using Slipstream.Components.Twitch;
 using Slipstream.Shared;
-using Slipstream.Shared.Factories;
 
 #nullable enable
 
-namespace Slipstream.Backend.Services.LuaServiceLib
+namespace Slipstream.Components.Twitch
 {
-    public class TwitchMethodCollection
+    public class LuaGlue : ILuaGlue
     {
         private readonly IEventBus EventBus;
         private readonly ITwitchEventFactory EventFactory;
 
-        public static TwitchMethodCollection Register(IEventBus eventBus, ITwitchEventFactory eventFactory, Lua lua)
-        {
-            var m = new TwitchMethodCollection(eventBus, eventFactory);
-            m.Register(lua);
-            return m;
-        }
-
-        public TwitchMethodCollection(IEventBus eventBus, ITwitchEventFactory eventFactory)
+        public LuaGlue(IEventBus eventBus, ITwitchEventFactory eventFactory)
         {
             EventBus = eventBus;
             EventFactory = eventFactory;
         }
 
-        public void Register(Lua lua)
+        public void SetupLua(NLua.Lua lua)
         {
             lua["twitch"] = this;
             lua.DoString(@"
