+++ Shared/EventFactory.cs
@@ -11,7 +11,6 @@ namespace Slipstream.Shared
         public EventFactory()
         {
             Factories.Add(typeof(IInternalEventFactory), new InternalEventFactory());
-            Factories.Add(typeof(ITwitchEventFactory), new TwitchEventFactory());
             Factories.Add(typeof(IUIEventFactory), new UIEventFactory());
         }
