+++ Shared/EventHandlerControllerBuilder.cs
@@ -14,7 +14,6 @@ namespace Slipstream.Shared
         {
             Add(typeof(Internal));
             Add(typeof(UI));
-            Add(typeof(Twitch));
         }
 
         public void Add(Type t)
